+++ Examples/Mac Example/FXKeychainExampleMac/AppDelegate.m
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.m
+//  FXKeychainExampleMac
+//
+//  Created by Nick Lockwood on 30/12/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import "AppDelegate.h"
+
+@implementation AppDelegate
+
+- (void)applicationDidFinishLaunching:(NSNotification *)aNotification
+{
+    // Insert code here to initialize your application
+}
+
+@end