+++ Examples/Mac Example/FXKeychainExampleMac/WindowController.h
@@ -1,0 +1,13 @@
+//
+//  WindowController.h
+//  FXKeychainExampleMac
+//
+//  Created by Nick Lockwood on 30/12/2012.
+//  Copyright (c) 2012 Charcoal Design. All rights reserved.
+//
+
+#import <Cocoa/Cocoa.h>
+
+@interface WindowController : NSWindowController
+
+@end