+++ RELEASE NOTES.md
@@ -1,3 +1,10 @@
+Version 1.1
+
+- Now uses application bundle ID to namespace the default keychain
+- Now supports keyed subcripting (e.g. keychain["foo"] = bar;)
+- Included CocoaPods podspec file
+- Incuded Mac example
+
 Version 1.0
 
 - Initial release